+++ truck-stepio/tests/io/main.rs
@@ -1,2 +1,3 @@
 mod ioi;
+mod modeling;
 mod oi;